% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviationr.R
\docType{data}
\name{x}
\alias{x}
\title{Standard deviation of the values in R}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
x
}
\arguments{
\item{sqrt}{to calculate Square root}

\item{sum}{to calculate sum}

\item{mean}{to calculate mean}
}
\value{
standard deviation
}
\description{
Calculates Standard deviation
}
\examples{
deviationr(x)

}
\keyword{datasets}
